Add ability to import recipe images (#19)
This PR adds the ability for the import feature to automatically download a recipe's image, offering a hotlinked preview in the import page. We also better handle imported recipes that have redundant names that end in ellipses. For example:

```yaml
name: First, preheat your oven to...
text: First, preheat your oven to 350F and get your ingredients ready
```

I think this is an SEO thing or a default behavior of Yoast, but it's unnecessary here so we get rid of it.

Could be an issue with valid cases like:

```yaml
name: Preheat
text: Preheat your oven to 350
```

Maybe we just act on ellipses names?