Import recipes with a single image and await deduped resizes (#24)
This PR improves the recipe import feature by properly importing recipe LD JSON with an `image` property that is a single object instead of an array. There is a new import test with a new fixture for this feature.

While working with the tests, I noticed images were being flaky. A tweak has been made to `next-static-image` so that when it re-uses a running resize, that reused resize is now awaited. This seems to fix the image flakiness.