Add "Seeking" feature to search (#34)
This PR modifies the search feature to keep loading pages until it finds an item, called "seeking". Seeking starts when a search is initiated, as well as when the "Load More" button is pressed. Once a new item is found, seeking stops and the "Load More" button can be used to initiate seeking again if more results are desired.